chapter 11 covered

+++ src/main/java/com/erfanlibrary/api/repository/BookRepository.java
@@ -1,10 +1,9 @@
 package com.erfanlibrary.api.repository;
 
 import com.erfanlibrary.api.model.Book;
-import com.erfanlibrary.api.model.CustomerBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
     Book findOneById(Integer id);
 
 }
